move dates to same paragraph as employer/degree
Since I'm not using the <time> tag for the dates because I want the
months to be read as text instead of numbers, I moved them to the same
paragraph as the one that contains my employer or degree earned. I'm
separating the date from the employer or degree with an em dash.