add Gradientify jQuery plugin
I wanted my background to have changing gradients, and I found a cool
jQuery plugin to do just that. I added the jquery.gradientify.min.js
file to the js folder, referenced it in the HTML and added the code to
the main.js file. Neat. Now to change the RGB values!